Removed AddOnSubscribeSingle, AddOnSubscribe now implements SingleOnSubscribe too
