+++ rxfirestore/src/main/java/com/oakwoodsc/rxfirestore/AddOnSubscribe.java
@@ -1,20 +1,20 @@
 package com.oakwoodsc.rxfirestore;
 
 import android.support.annotation.NonNull;
-
 import com.google.android.gms.tasks.OnCompleteListener;
 import com.google.android.gms.tasks.Task;
 import com.google.firebase.firestore.CollectionReference;
 import com.google.firebase.firestore.DocumentReference;
-
 import io.reactivex.CompletableEmitter;
 import io.reactivex.CompletableOnSubscribe;
+import io.reactivex.SingleEmitter;
+import io.reactivex.SingleOnSubscribe;
 
 /**
  * Created by Brandon Trautmann
  */
 
-public class AddOnSubscribe<T> implements CompletableOnSubscribe {
+public class AddOnSubscribe<T> implements CompletableOnSubscribe, SingleOnSubscribe<DocumentReference> {
 
   private final CollectionReference reference;
   private final T value;
@@ -39,6 +39,25 @@ public class AddOnSubscribe<T> implements CompletableOnSubscribe {
               }
             }
 
+          }
+        };
+
+    reference.add(value).addOnCompleteListener(listener);
+  }
+
+  @Override
+  public void subscribe(final SingleEmitter<DocumentReference> emitter) throws Exception {
+    final OnCompleteListener<DocumentReference> listener =
+        new OnCompleteListener<DocumentReference>() {
+          @Override
+          public void onComplete(@NonNull Task<DocumentReference> task) {
+            if (!emitter.isDisposed()) {
+              if (!task.isSuccessful()) {
+                emitter.onError(task.getException());
+              } else {
+                emitter.onSuccess(task.getResult());
+              }
+            }
           }
         };
